@@ -9,8 +9,7 @@ namespace kalanis\kw_mapper\Mappers\Database;
  */
 trait TTable
 {
-    /** @var string */
-    protected $tableName = '';
+    protected string $tableName = '';
 
     public function setTable(string $tableName): void
     {
